Split each resolver to different files
