+++ lib/resolvers/task_resolver.ex
@@ -1,0 +1,34 @@
+defmodule Gannbaruzoi.TaskResolver do
+  alias Gannbaruzoi.Repo
+  alias Gannbaruzoi.Task
+
+  def all(_, _) do
+     # TODO: scope with user_id
+     task = Repo.all(Task)
+     {:ok, task}
+  end
+
+  def create(_parent, attributes, _info) do
+    changeset = Task.changeset(%Task{type: "root"}, attributes)
+
+    case Repo.insert(changeset) do
+      {:ok, task} -> {:ok, %{task: task}}
+      {:error, changeset} -> {:error, changeset.errors}
+    end
+  end
+
+  def update(_parent, attributes, _info) do
+    changeset = Repo.get!(Task, attributes.id) |> Task.changeset(attributes)
+
+    case Repo.update(changeset) do
+      {:ok, task} -> {:ok, %{task: task}}
+      {:error, changeset} -> {:error, changeset.errors}
+    end
+  end
+
+  def delete(_parent, attributes, _info) do
+    task = Repo.get!(Task, attributes.id)
+    Repo.delete(task)
+    {:ok, %{id: task.id}}
+  end
+end
